Rename transition scheme path: t8_schemes/t8_transition_quad/.. to t8_schemes/t8_transition/t8_transition_conformal_quad, such that further transition schemes can easily be added.
