+++ Makefile.am
@@ -73,7 +73,7 @@ DISTCLEANFILES += \
 # non-recursive build
 include src/Makefile.am
 include src/t8_schemes/t8_default/Makefile.am
-include src/t8_schemes/t8_quads_transition/t8_transition/Makefile.am
+include src/t8_schemes/t8_transition/t8_transition_conformal_quad/Makefile.am
 include test/Makefile.am
 include example/Makefile.am
 include tutorials/Makefile.am
