--- src/t8_schemes/t8_quads_transition/t8_transition/Makefile.am
@@ -1,12 +1,0 @@
-# This file is part of t8code
-# Non-recursive Makefile.am in src/t8_schemes/t8_quads_transition
-# Included from toplevel directory
-
-libt8_installed_headers += \
-  src/t8_schemes/t8_quads_transition/t8_transition_cxx.hxx \
-  src/t8_schemes/t8_quads_transition/t8_transition/t8_transition_quad_cxx.hxx 
-
-libt8_compiled_sources += \
-  src/t8_schemes/t8_quads_transition/t8_transition/t8_transition_cxx.cxx \
-  src/t8_schemes/t8_quads_transition/t8_transition/t8_transition_quad_cxx.cxx
-
