+++ src/t8_schemes/t8_transition/t8_transition_conformal_quad/t8_transition_conformal_cxx.cxx
@@ -21,11 +21,13 @@
 */
 
 #include <new>
-#include <t8_schemes/t8_quads_transition/t8_transition_cxx.hxx>
+#include <t8_schemes/t8_transition/t8_transition_cxx.hxx>
 #include <t8_schemes/t8_default/t8_default_line/t8_default_line_cxx.hxx>
+/* TODO: Note that, if a transition_conformal version of triangles would be implemented,
+ * then this version could be included here, enabling hybrid-quad-triangle transitioned conformal meshes. */
 #include <t8_schemes/t8_default/t8_default_tri/t8_default_tri_cxx.hxx>
 #include <t8_refcount.h>
-#include "t8_transition_quad_cxx.hxx"
+#include "t8_transition_conformal_quad_cxx.hxx"
 
 /* We want to export the whole implementation to be callable from "C" */
 T8_EXTERN_C_BEGIN ();
@@ -39,7 +41,7 @@ t8_scheme_new_subelement_cxx (void)
   t8_refcount_init (&s->rc);
 
   s->eclass_schemes[T8_ECLASS_VERTEX] = NULL;
-  s->eclass_schemes[T8_ECLASS_LINE] = new t8_default_scheme_line_c ();
+  s->eclass_schemes[T8_ECLASS_LINE] = new t8_default_scheme_line_c (); /* Standard T8_ECLASS_LINE is used by the quad scheme */
   s->eclass_schemes[T8_ECLASS_QUAD] = new t8_subelement_scheme_quad_c ();
   s->eclass_schemes[T8_ECLASS_HEX] = NULL;
   s->eclass_schemes[T8_ECLASS_TRIANGLE] = new t8_default_scheme_tri_c (); /* we use this implementation for hybrid meshes */
