+++ src/t8_schemes/t8_transition/t8_transition_conformal_quad/t8_transition_conformal_quad_cxx.cxx
@@ -27,7 +27,7 @@
 #include <t8_schemes/t8_default/t8_default_line/t8_dline_bits.h>
 #include <t8_schemes/t8_default/t8_default_common/t8_default_common_cxx.hxx>
 #include "t8.h"
-#include "t8_transition_quad_cxx.hxx"
+#include "t8_transition_conformal_quad_cxx.hxx"
 
 /* *INDENT-OFF* */
 /* Connectivity of subelement faces:
